增加Swagger-ui

+++ model/src/main/java/com/sean/model/repo/MemberRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sean.model.entities.MemberEntity;
 
-public interface MemberRepo extends JpaRepository<MemberEntity, Long> {
+public interface MemberRepo extends JpaRepository<MemberEntity, Integer> {
 
 }
